Refactor strategy merging and fallback handling
